remove dead file index.tsx
