+++ wormhole-connect/src/main.tsx
@@ -1,9 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
+import ErrorBoundary from './components/ErrorBoundary';
 
-ReactDOM.createRoot(document.getElementById('wormhole-connect')!).render(
+export * from './theme';
+
+const root = ReactDOM.createRoot(
+  document.getElementById('wormhole-connect') as HTMLElement,
+);
+root.render(
   <React.StrictMode>
-    <App />
+    <ErrorBoundary>
+      <App />
+    </ErrorBoundary>
   </React.StrictMode>,
 );
